backend/requirements.txt
# ============================================================================================
# 📦 requirements.txt – Dependências do backend Agroverso
# --------------------------------------------------------------------------------------------
# 🎯 Inclui FastAPI, segurança JWT, ORM, criptografia e adaptador serverless
# ✨ Versões testadas, compatíveis com Vercel Functions e ambientes Python 3.10+
# ============================================================================================

fastapi==0.110.0
uvicorn==0.29.0
mangum==0.17.0

sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1

python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# (opcional para envio de e-mails, uploads ou cache futuro)
# httpx==0.27.0
# sendgrid==6.11.0
# redis==5.0.3
